+++ src/main/java/com/kakaotech/team14backend/jwt/Controller/JwtController.java
@@ -1,18 +1,11 @@
 package com.kakaotech.team14backend.jwt.Controller;
 
 
-import com.auth0.jwt.exceptions.TokenExpiredException;
-import com.kakaotech.team14backend.auth.PrincipalDetails;
 import com.kakaotech.team14backend.common.ApiResponse;
-import com.kakaotech.team14backend.common.ApiResponse.CustomBody;
 import com.kakaotech.team14backend.common.ApiResponseGenerator;
 import com.kakaotech.team14backend.common.CookieUtils;
-import com.kakaotech.team14backend.exception.TokenValidationException;
-import com.kakaotech.team14backend.filter.FilterResponseUtils;
 import com.kakaotech.team14backend.jwt.dto.ReissueDTO;
-import com.kakaotech.team14backend.jwt.dto.TokenDTO;
 import com.kakaotech.team14backend.jwt.service.TokenService;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
