+++ src/main/java/com/kakaotech/team14backend/jwt/service/TokenService.java
@@ -5,14 +5,14 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.kakaotech.team14backend.common.MessageCode;
-import com.kakaotech.team14backend.exception.MemberNotFoundException;
-import com.kakaotech.team14backend.exception.TokenValidationException;
-import com.kakaotech.team14backend.member.domain.Member;
-import com.kakaotech.team14backend.member.infrastructure.MemberRepository;
 import com.kakaotech.team14backend.jwt.RefreshToken;
+import com.kakaotech.team14backend.jwt.TokenValidationException;
 import com.kakaotech.team14backend.jwt.dto.ReissueDTO;
 import com.kakaotech.team14backend.jwt.dto.TokenDTO;
 import com.kakaotech.team14backend.jwt.repository.RefreshTokenRepository;
+import com.kakaotech.team14backend.member.domain.Member;
+import com.kakaotech.team14backend.member.exception.MemberNotFoundException;
+import com.kakaotech.team14backend.member.infrastructure.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
